requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.1.2
fire==0.3.1
idna==2.8
jieba==0.42.1
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
nltk==3.5
numpy==1.19.2
Pillow==7.2.0
PyMySQL==0.10.1
pyparsing==2.4.7
PyQt5==5.13.2
PyQt5-sip==12.8.1
pyqt5-tools==5.13.2.1.6rc1
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2017.4.5
requests==2.21.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.3
wxPython==4.1.0
